+++ src/types/Config.tsx
@@ -82,6 +82,7 @@ import AddRoles from 'components/pages/AppOverview/AddRoles'
 import ServiceDeactivate from 'components/pages/ServiceReleaseProcess/components/ServiceDeactivate'
 import ChangeDocuments from 'components/pages/AppOverview/ChangeDocuments'
 import OSPManagement from 'components/pages/OSPManagement'
+import CompanyCertificates from 'components/pages/CompanyCertificates'
 
 /**
  * ALL_PAGES
@@ -532,6 +533,11 @@ export const ALL_PAGES: IPage[] = [
     role: ROLES.IDP_VIEW,
     element: <OSPManagement />,
   },
+  {
+    name: PAGES.COMPANY_CERTIFICATE,
+    role: ROLES.COMPANY_CERTIFICATE_VIEW,
+    element: <CompanyCertificates />,
+  }
 ]
 
 export const ALL_OVERLAYS: IOverlay[] = [
@@ -769,6 +775,7 @@ export const userMenuFull = [
   PAGES.USECASE_PARTICIPATION,
   PAGES.CERTIFICATE_CREDENTIAL,
   PAGES.ADMIN_CREDENTIAL,
+  PAGES.COMPANY_CERTIFICATE,
   PAGES.LOGOUT,
 ]