+++ src/types/Constants.ts
@@ -101,6 +101,7 @@ export enum PAGES {
   DATA_SPACE = 'dataspace',
   ADMIN_CREDENTIAL = 'admin-credential',
   ONBOARDING_SERVICEPROVIDER = 'onboarding-serviceprovider',
+  COMPANY_CERTIFICATE = 'companyCertificate'
 }
 
 export enum OVERLAYS {
@@ -213,6 +214,8 @@ export enum ROLES {
   SUBMITTED_APPLICATION = 'view_submitted_applications',
   REQUEST_SSICREDENTIAL = 'request_ssicredential',
   DECISION_SSICREDENTIAL = 'decision_ssicredential',
+  COMPANY_CERTIFICATE_VIEW = 'view_certificates',
+  UPLOAD_COMPANY_CERTIFICATE = 'upload_certificates'
 }
 
 export enum HINTS {